feat: :sparkles: Team populated model 제시, 필수 파일구조 구현
로직은 전혀 없는 껍데기 상태입니다. 추후 구현해야 하고, xxxPopulated 타입들도 구현하는대로 업데이트해야 합니다.

- #5
