+++ app/src/teams/models/team.model.ts
@@ -51,4 +51,10 @@ export class Team {
 
   @Field((_type) => [TeamScaleTeam], { nullable: 'items' })
   teamScaleTeams: TeamScaleTeam[];
+
+  @Field()
+  usersPopulated: string; // todo: user
+
+  @Field()
+  projectPopulated: string; // todo: project
 }
