+++ app/src/teams/models/team.populated.model.ts
@@ -1,0 +1,8 @@
+import { ObjectType, OmitType } from '@nestjs/graphql';
+import { Team } from './team.model';
+
+@ObjectType()
+export class TeamPopulated extends OmitType(Team, [
+  'usersPopulated',
+  'projectPopulated',
+]) {}
